+++ web/next/components/blog/modal/OutlineModal.module.scss
